refactor: index.js 코드 간결화
